[CWS] move reorder to own module
