+++ pkg/security/probe/reorderer/reorderer_monitor.go
@@ -6,25 +6,27 @@
 //go:build linux
 // +build linux
 
-package probe
+package reorderer
 
 import (
 	"context"
 	"sync"
 
 	"github.com/DataDog/datadog-agent/pkg/security/metrics"
+	"github.com/DataDog/datadog-go/v5/statsd"
 )
 
 // ReordererMonitor represents a reorderer monitor
 type ReordererMonitor struct {
-	// probe is a pointer to the Probe
-	probe *Probe
+	monitored    *ReOrderer
+	statsdClient statsd.ClientInterface
 }
 
 // NewReOrderMonitor instantiates a new reorder statistics counter
-func NewReOrderMonitor(p *Probe) (*ReordererMonitor, error) {
+func NewReOrderMonitor(r *ReOrderer, statsdClient statsd.ClientInterface) (*ReordererMonitor, error) {
 	return &ReordererMonitor{
-		probe: p,
+		monitored:    r,
+		statsdClient: statsdClient,
 	}, nil
 }
 
@@ -34,13 +36,13 @@ func (r *ReordererMonitor) Start(ctx context.Context, wg *sync.WaitGroup) {
 
 	for {
 		select {
-		case metric := <-r.probe.reOrderer.Metrics:
-			_ = r.probe.statsdClient.Gauge(metrics.MetricPerfBufferSortingQueueSize, float64(metric.QueueSize), []string{}, 1.0)
+		case metric := <-r.monitored.Metrics:
+			_ = r.statsdClient.Gauge(metrics.MetricPerfBufferSortingQueueSize, float64(metric.QueueSize), []string{}, 1.0)
 			var avg float64
 			if metric.TotalOp > 0 {
 				avg = float64(metric.TotalDepth) / float64(metric.TotalOp)
 			}
-			_ = r.probe.statsdClient.Gauge(metrics.MetricPerfBufferSortingAvgOp, avg, []string{}, 1.0)
+			_ = r.statsdClient.Gauge(metrics.MetricPerfBufferSortingAvgOp, avg, []string{}, 1.0)
 		case <-ctx.Done():
 			return
 		}
